Add new sparkline printer[1] which uses the jquery.sparkline.js[2]
## Description
The plugin generates sparklines (small inline charts) to be displayed
within a text.

Data are retrievd from class SMWAggregatablePrinter

It still uses the old getParameters since I'm having trouble
with the new getParamDefinitions

## Note
[1] http://www.semantic-mediawiki.org/wiki/Help:Sparkline_format
[2] http://omnipotent.net/jquery.sparkline/ and is published under
the New BSD License

Plugin is compatible with most modern browsers and has been
tested with Firefox 2+, Safari 3+, Opera 9, Google Chrome and
Internet Explorer 6, 7, 8 & 9 as well as iOS and Android.

## Patchsets
+1 initial commit, added test file, jquery files,
+2 fixed $class of undefined
+3 fixed itendation
+4 something happened with the jquery during transfer, hope this
fixed it
+5 #4 didn't solve the problem, I'm using a non minified file now
+6 #5 jquery.sparkline.js works as advertised, this patchset is
just to verfify the status of jquery.sparkline.min.js
+7 uploaded new minified version from the website and apparently
it works therefore we stick with the minified file
(deletd non-minified)
+8 Changed from getParameters() to getParamDefinitions()
+9 Changed test class name (SparklineTest) and updated release notes

Change-Id: I911862cedcb13aed77f7765bc138cb1f9f897a30