SRF event calendar add option to determine an initial calendar start
## Description
Add an option to determine an initial calendar start. Use "earliest"
for selected events instead of assuming the current year/month/day.

"earliest" use case
For a specifc group of planned events (outside of the current month)
setting the calendar display of the ealieast event selected and
making it possible to show scheduled events beyond the current month
for the initial calendar period.

+ Change the way specific properties are used, identify properties
through a fixed identifier assigned to a property. This allows a
property to be variable within a query while its identifier
reamain fixed without the need for extra parameters (Fixed identifiers
are title, icon, color) resulting in
...
|?has event icon=icon
|?has color=color
|?event=title
...

[1] http://www.semantic-mediawiki.org/wiki/Event_calendar

Change-Id: I9e7ecd60563a97e54ad2952634f368275d4d7ab3