bugfix (timeline loaded partially)
The Simile timeline module often only loads partially since the ResourceLoader was introduced.
Reason seems to be, that it loads its files by appending script tags to the head which is
already read by the browser when the ResourceLoader loads modules. Requesting a 'top' position
for the module seems to fix this.

Second change is the removal of a deprecated function call and replacing it by a jQuery event
hook registration.

Change-Id: I656b14198c94da3b9b9bc6e2b587247855faa572