Update SemanticResultFormats.php
If another composer-installed extension (e.g. Semantic Extra Special Properties) installs Semantic MediaWiki in its local extension vendor directory, Semantic MediaWiki will be autoloaded from there, but the files for this extension will never be autoloaded (because SMW_VERSION will already be defined). Perhaps there is a better solution, but with this change, if composer is run in the SemanticResultFormats directory to create its local vendor directory, the code will now be sure to autoload the files for this extension.