Add option to use DEFAULTSORT in Filtered format (#225)
When using the Filtered result format, a field specified as a value filter when using can have its ordering specified explicitly using the "value filter values" parameter, or it can use a default behavior of alphabetically sorting based on the value of that field. However, neither of these behaviors are particularly useful when the filter is on a field where the values are page names (for example, something like Person:123): normal alphabetization would put them in a strange order (where Person:123 precedes Person:2, for instance), and when an explicit order listing (which would have to be generated by some other function) is used, the same value is used both for the checkbox label as for the sought value in the field (so a checkbox that was labeled "Thompson, Matthew" would filter based on that string, which would match no results, as the value in the result is e.g. "Person:123").

A desirable behavior in this case would be to use the page's DEFAULTSORT attribute to establish the ordering. To this end, additional sort information is generated based on the DEFAULTSORT attribute of the pages that are found when using the field's value as a page title. When this information is present in a field that is being used to create filter checkboxes, the sort is applied, and the checkboxes appear in the order specified by DEFAULTSORT.