Add includeend parameter for eventcalendar format
When the includeend parameter is set to true, the end date is
included in the date range for an event. This is useful on wikis
where event start and end dates, but not times, are entered. Since
the default time is midnight, the entered end date would normally
not be included in the date range.

The desired output is achieved by adding one day to the end date
before setting the rowData. This is similar to how the eventcalendar
library implements its nextDayThreshold functionality.