Use orthogonal data for DataTables (#865)
* Use orthogonal data for DataTables

Instead of assigning HTML directly as data, we can use assign an object with the following keys:

- display - The display value of the data (same as what we had)

- filter - Value used for filtering (e.g. SearchBuilder)

- order - Value used for sorting

This will make DataTables easier to expand in the long run. Since we can assign different

attributes to the data.

Also fixes #831

* Drop any-number from DataTables

We don't need it anymore because values are assigned a sorting key

* Remove datatables-columns.type as we do not need it anymore

* Use sort instead of order for key name

* Ensure that dataValues[0] exists in SortKey